+++ custom_components/sber_mqtt_bridge/devices/climate.py
@@ -236,4 +236,3 @@ class ClimateEntity(BaseEntity):
                         }
                     )
         return results
-