+++ custom_components/sber_mqtt_bridge/devices/curtain.py
@@ -208,4 +208,3 @@ class CurtainEntity(BaseEntity):
         )
 
         return {self.entity_id: {"states": states}}
-